src/news_section/requirements.txt
pandas
beautifulsoup4
httpx
yfinance
finvizfinance
praw
pyarrow
python-dotenv
google-cloud-bigquery
